chore: split base from webpack.config.js
