+++ webpack.config.js
@@ -1,40 +1,6 @@
-const path = require('path');
-// const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
-const webpack = require('webpack');
+const baseConfig = require('./webpack.base.config');
 const nodeExternals = require('webpack-node-externals');
 
-const baseConfig = {
-  entry: './src/index.ts',
-  // devtool: 'inline-source-map',
-  devtool: 'cheap-module-source-map',
-  // mode: 'development',
-  mode: "production",
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        use: 'ts-loader',
-        exclude: /node_modules/,
-      },
-    ],
-  },
-  node: {
-    // Prevent embedded winston to throw error with FS not existing.
-    fs: 'empty',
-  },
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-  },
-  resolve: {
-    extensions: ['.ts', '.js', '.json'],
-  },
-  plugins: [
-    // new BundleAnalyzerPlugin(),
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify('production')
-    })
-  ],
-}
 const webConfig = Object.assign({}, baseConfig, {
   target: 'web',
   output: {
